Update player prefab and added PlayerAttackController.cs.
