+++ Lucifer's Trials/Assets/Scripts/PlayerAttackController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a23459d879e3a7aa390a35869706b17f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
